+++ src/tipl/ij/volviewer/TextField.java
@@ -12,9 +12,9 @@ class TextField {
 	private String text = "";
 	private Color color;
 
-	private float xpos;
-	private float ypos;
-	private float z;
+	private double xpos;
+	private double ypos;
+	private double z;
 	private int fontNr;
 
 	public TextField(String text, Color color, int xpos, int ypos, int z) {
@@ -46,11 +46,11 @@ class TextField {
 		this.text = text;
 	}
 
-	public void setXpos(float xpos) {
+	public void setXpos(double xpos) {
 		this.xpos = xpos;
 	}
 
-	public void setZ(float z) {
+	public void setZ(double z) {
 		this.z = z;
 	}
 
@@ -58,7 +58,7 @@ class TextField {
 		return color;
 	}
 
-	public float getZ() {
+	public double getZ() {
 		return z;
 	}
 
@@ -70,7 +70,7 @@ class TextField {
 		this.fontNr = fontNr;
 	}
 
-	public float getXpos() {
+	public double getXpos() {
 		return xpos;
 	}
 
@@ -78,11 +78,11 @@ class TextField {
 		return text;
 	}
 
-	public float getYpos() {
+	public double getYpos() {
 		return ypos;
 	}
 
-	public void setYpos(float ypos) {
+	public void setYpos(double ypos) {
 		this.ypos = ypos;
 	}
 }
